add extra quartus commands in check_synthesizable.py
Also add the extra quartus commands to get the cva5 to build.

Basically some import <pkg>::* statements are missing in the cva5
source.

In Vivado's interactive mode (which is where the cva5 was developed) .sv source
is bundled together and then synthesized, so as long as packages appear further
up in the bundle than the first use of symbols they define everything works
even with the missing imports. Further, Vivado is smart enough to figure out
the correct ordering to bundle things (update_compile_order).

By default Quartus does not bundle .sv source and so missing imports are
an error. You can tell Quartus to bundle source (VERILOG_CU_MODE MFCU,
"multifile compile unit") but then its up to you to enforce correct
ordering in the bundle by adding source files in the right order. The
tcl in quartus-extra-commands adds all of the packages in the correct
dependancy order, and then adds all of the non-package source files.